requirements.txt
python-telegram-bot==20.6
Flask==2.2.5
waitress==2.1.2
